requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
line-bot-sdk==2.4.2
gunicorn==23.0.0
psycopg2-binary==2.9.9
